+++ src/rpc/mining.cpp
@@ -924,36 +924,6 @@ static UniValue estimatefee(const Config &config,
     return ValueFromAmount(feeRate.GetFeePerK());
 }
 
-static UniValue estimatepriority(const Config &config,
-                                 const JSONRPCRequest &request) {
-    if (request.fHelp || request.params.size() != 1) {
-        throw std::runtime_error(
-            "estimatepriority nblocks\n"
-            "\nDEPRECATED. Estimates the approximate priority "
-            "a zero-fee transaction needs to begin\n"
-            "confirmation within nblocks blocks.\n"
-            "\nArguments:\n"
-            "1. nblocks     (numeric, required)\n"
-            "\nResult:\n"
-            "n              (numeric) estimated priority\n"
-            "\n"
-            "A negative value is returned if not enough "
-            "transactions and blocks\n"
-            "have been observed to make an estimate.\n"
-            "\nExample:\n" +
-            HelpExampleCli("estimatepriority", "6"));
-    }
-
-    RPCTypeCheck(request.params, {UniValue::VNUM});
-
-    int nBlocks = request.params[0].get_int();
-    if (nBlocks < 1) {
-        nBlocks = 1;
-    }
-
-    return mempool.estimatePriority(nBlocks);
-}
-
 static UniValue estimatesmartfee(const Config &config,
                                  const JSONRPCRequest &request) {
     if (request.fHelp || request.params.size() != 1) {
@@ -1054,7 +1024,6 @@ static const CRPCCommand commands[] = {
     {"generating", "generatetoaddress",     generatetoaddress,     true, {"nblocks", "address", "maxtries"}},
 
     {"util",       "estimatefee",           estimatefee,           true, {"nblocks"}},
-    {"util",       "estimatepriority",      estimatepriority,      true, {"nblocks"}},
     {"util",       "estimatesmartfee",      estimatesmartfee,      true, {"nblocks"}},
     {"util",       "estimatesmartpriority", estimatesmartpriority, true, {"nblocks"}},
 };